Cleanup and BgClipCheckOpeningHatch fix

+++ include/bg_clip.h
@@ -18,7 +18,7 @@ void BgClipCheckTouchingTransitionOrTank(void);
 void BgClipFinishCollectingTank(void);
 void BgClipFinishCollectingAbility(void);
 void BgClipCheckGrabbingCrumbleBlock(u8 dontDestroy);
-u8 BgClipCheckOpeningHatch(u16 xPosition, u16 yPosition);
+u32 BgClipCheckOpeningHatch(u16 xPosition, u16 yPosition);
 void BgClipSetItemAsCollected(u16 xPosition, u16 yPosition, u8 type);
 void BgClipRemoveCollectedTanks(void);
 void BgClipCallMotherBrainUpdateGlass(u8 stage);
